pyproject.toml
[project]
name = "doc2embed"
version = "0.1.0"
description = "FastAPI service for doc2embed"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0,<1.0.0",
    "python-docx>=1.1.2,<2.0.0",
    "pydantic-settings>=2.6.0,<3.0.0",
    "psycopg[binary]>=3.2.0,<4.0.0",
    "sqlalchemy>=2.0.0,<3.0.0",
    "uvicorn[standard]>=0.30.0,<1.0.0",
]

[project.scripts]
doc2embed = "main:run"
